Moved submit_chunk function from rotsearch to util as it is a general function
