+++ simbad/rotsearch/__init__.py
@@ -4,9 +4,6 @@ __author__ = "Adam Simpkin & Felix Simkovic"
 __date__ = "12 April 2018"
 __version__ = "0.3"
 
-import math
-from pyjob.factory import TaskFactory
-
 
 def rotation_search_factory(method):
     if method == "amore":
@@ -34,47 +31,4 @@ def get_total_chunk_cycles(total, step):
         return total_chunk_cycles + 1
     else:
         return total_chunk_cycles
-
-
-def submit_chunk(collector, run_dir, nproc, job_name, submit_qtype, submit_queue, monitor, success_func):
-    """Submit jobs in small chunks to avoid using too much disk space
-
-    Parameters
-    ----------
-    collector : list
-        :obj:`~pyjob.script.ScriptCollector` containing run scripts
-    nproc : int, optional
-        The number of processors to run the job on
-    job_name : str
-        The name of the job to submit
-    submit_qtype : str
-        The cluster submission queue type - currently support SGE and LSF
-    submit_queue : str
-        The queue to submit to on the cluster
-    success_func : func
-        function to check for success
-
-    """
-
-    if submit_qtype == 'local':
-        processes = nproc
-        array_size = None
-    else:
-        processes = None
-        array_size = nproc
-
-    with TaskFactory(submit_qtype,
-                     collector,
-                     cwd=run_dir,
-                     name=job_name,
-                     processes=processes,
-                     max_array_size=array_size,
-                     queue=submit_queue,
-                     permit_nonzero=True,
-                     shell='/bin/bash',
-                     priority=-10) as task:
-        task.run()
-        interval = int(math.log(len(collector.scripts)) / 3)
-        interval_in_seconds = interval if interval >= 5 else 5
-        task.wait(interval=interval_in_seconds, monitor_f=monitor, success_f=success_func)
